Agurment and type refactore

+++ principles/src/main/java/com/thecollinsbyte/principles/SOLID/openclosed/Calculator.java
@@ -1,7 +1,7 @@
 package com.thecollinsbyte.principles.SOLID.openclosed;
 
 public class Calculator {
-    public int calculateNumber(int number1, int number2, Operation operation) {
+    public double calculateNumber(double number1, double number2, Operation operation) {
         return operation.calculate(number1, number2);
     }
 }
