+++ principles/src/main/java/com/thecollinsbyte/principles/SOLID/openclosed/SubtractOperation.java
@@ -3,7 +3,7 @@ package com.thecollinsbyte.principles.SOLID.openclosed;
 public class SubtractOperation implements Operation {
 
     @Override
-    public int calculate(int num1, int num2) {
+    public double calculate(double num1, double num2) {
         return num1 - num2;
     }
 }
